refactor(frontend): extract NavSearch from nav (571 -> 343 lines)
